♻️ Refactor: 복잡한 네이밍 수정 및 디렉토리 분리
